Improved signup process
