+++ webapp/i18n/en.json
@@ -1489,6 +1489,12 @@
   "sidebar_right_menu.switch_team": "Team Selection",
   "sidebar_right_menu.teamLink": "Get Team Invite Link",
   "sidebar_right_menu.teamSettings": "Team Settings",
+  "signup.title": "Create an account with:",
+  "signup.ldap": "LDAP Credentials",
+  "signup.google": "Google Account",
+  "signup.office365": "Office 365",
+  "signup.gitlab": "GitLab Single-Sign-On",
+  "signup.email": "Email and Password",
   "signup_team.choose": "Your teams: ",
   "signup_team.createTeam": "Or Create a Team",
   "signup_team.disabled": "Team creation has been disabled.  Please contact an administrator for access.",
