+++ webapp/sass/routes/_signup.scss
@@ -203,6 +203,11 @@
         .fa {
             font-size: 17px;
             margin-right: 8px;
+
+            &.fa--margin-top {
+                position: relative;
+                top: 2px;
+            }
         }
 
         .icon {
@@ -210,6 +215,7 @@
             display: inline-block;
             height: 18px;
             margin-right: 8px;
+            text-align: center;
             width: 18px;
         }
 
@@ -218,10 +224,11 @@
             color: $white;
             display: block;
             height: 40px;
-            line-height: 34px;
+            line-height: 36px;
             margin: 1em 0;
             min-width: 200px;
-            padding: 0 1em;
+            padding: 0 1em 0 2em;
+            text-align: left;
             width: 200px;
 
             &.gitlab {
@@ -268,16 +275,15 @@
                 }
 
                 .icon {
-                    margin-left:-10px;
                     background-image: url('../images/office365Logo.png');
                 }
             }
 
             &.ldap {
-                background: #dd4b39;
+                background: #3AA1CF;
 
                 &:hover {
-                    background: darken(#dd4b39, 10%);
+                    background: darken(#3AA1CF, 10%);
                 }
 
                 span {
@@ -298,10 +304,10 @@
             }
 
             &.saml {
-                background: #dd4b39;
+                background: #34a28b;
 
                 &:hover {
-                    background: darken(#dd4b39, 10%);
+                    background: darken(#34a28b, 10%);
                 }
 
                 span {
@@ -309,10 +315,15 @@
                 }
             }
 
-            &.btn-full {
+            &.btn--full {
+                max-width: 350px;
                 padding-left: 35px;
                 text-align: left;
                 width: 100%;
+            }
+
+            &.btn--large {
+                width: 300px;
             }
         }
